move events to scm-core / move subscriber to landingpage-plugin
