+++ scm-webapp/pom.xml
@@ -24,7 +24,8 @@
     SOFTWARE.
 
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
   <modelVersion>4.0.0</modelVersion>
 
@@ -462,16 +463,6 @@
       <version>2.1.1</version>
       <scope>provided</scope>
     </dependency>
-
-
-    <!-- optional dependencies -->
-    <dependency>
-      <groupId>sonia.scm.plugins</groupId>
-      <artifactId>scm-landingpage-plugin</artifactId>
-      <version>1.0.0-SNAPSHOT</version>
-      <optional>true</optional>
-    </dependency>
-
   </dependencies>
 
   <build>
